Fix minor typos and consistency issues in the validation algorithm.
This patch fixes a couple of typos in the pseudocode listings for the
validation algorithm. It also introduces a small consistency change as
it normalises the signatures of `is_num`, `is_vec`, and `is_ref`,
i.e. this patch drops the the union type `val_type | Unknown` on the
input signatures of `is_vec` and `is_ref` in favour of simply
`val_type`. Consequently, the body of `is_vec` now checks whether the
input `t` is `Bot` in the right operand. I believe this change
reflects the original intent as the subtyping subsumes the `Unknown`
option.